Fix [Models] register model fails with 'Model failed to initiate'
